+++ apps/web/src/api-clients/inventory/collector/composables/use-collector-api.ts
@@ -9,7 +9,7 @@ import type { CollectorListParameters } from '@/api-clients/inventory/collector/
 import type { CollectorUpdateParameters } from '@/api-clients/inventory/collector/schema/api-verbs/update';
 import type { CollectorUpdatePluginParameters } from '@/api-clients/inventory/collector/schema/api-verbs/update-plugin';
 import type { CollectorModel } from '@/api-clients/inventory/collector/schema/model';
-import type { JobModel } from '@/schema/inventory/job/model';
+import type { JobModel } from '@/api-clients/inventory/job/schema/model';
 
 
 export const useCollectorApi = () => {
