+++ apps/web/src/api-clients/inventory/job/composables/use-job-api.ts
@@ -1,0 +1,17 @@
+import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
+
+import type { ListResponse } from '@/api-clients/_common/schema/api-verbs/list';
+import type { JobAnalyzeParameters } from '@/api-clients/inventory/job/schema/api-verbs/analyze';
+import type { JobListParameters } from '@/api-clients/inventory/job/schema/api-verbs/list';
+import type { JobModel } from '@/api-clients/inventory/job/schema/model';
+
+export const useInventoryJobApi = () => {
+    const actions = {
+        list: SpaceConnector.clientV2.inventory.job.list<JobListParameters, ListResponse<JobModel>>,
+        analyze: SpaceConnector.clientV2.inventory.job.analyze<JobAnalyzeParameters, ListResponse<JobModel>>,
+    };
+
+    return {
+        jobAPI: actions,
+    };
+};
