+++ apps/web/src/api-clients/inventory/job/schema/api-verbs/analyze.ts
