+++ apps/web/src/api-clients/inventory/job/schema/type.ts
