+++ apps/web/src/services/asset-inventory/stores/collector-data-modal-store.ts
@@ -7,9 +7,9 @@ import type { ListResponse } from '@/api-clients/_common/schema/api-verbs/list';
 import type {
     CollectorModel,
 } from '@/api-clients/inventory/collector/schema/model';
+import type { JobListParameters } from '@/api-clients/inventory/job/schema/api-verbs/list';
+import type { JobModel } from '@/api-clients/inventory/job/schema/model';
 import type { SecretModel } from '@/api-clients/secret/secret/schema/model';
-import type { JobListParameters } from '@/schema/inventory/job/api-verbs/list';
-import type { JobModel } from '@/schema/inventory/job/model';
 
 import ErrorHandler from '@/common/composables/error/errorHandler';
