+++ apps/web/src/services/asset-inventory/stores/collector-job-store.ts
@@ -7,8 +7,8 @@ import { ApiQueryHelper } from '@cloudforet/core-lib/space-connector/helper';
 import type { ListResponse } from '@/api-clients/_common/schema/api-verbs/list';
 import type { CollectorModel } from '@/api-clients/inventory/collector/schema/model';
 import type { Schedule } from '@/api-clients/inventory/collector/schema/type';
-import type { JobListParameters } from '@/schema/inventory/job/api-verbs/list';
-import type { JobModel } from '@/schema/inventory/job/model';
+import type { JobListParameters } from '@/api-clients/inventory/job/schema/api-verbs/list';
+import type { JobModel } from '@/api-clients/inventory/job/schema/model';
 
 import ErrorHandler from '@/common/composables/error/errorHandler';
